+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Changes in March 2026
 -   Updated `BubbleTreeArranger`, and `HierarchyArranger` constructor to pass `MutableIntList` instead of `Set<Integer>`. Old constructor has been deprecated.
+-   Updated `GraphSpectrumEmbedder.spectralEmbedding()` return type to `MutableIntObjectMap`, and `includedVertices` parameter to a `MutableIntSet`
 
 ## Changes in January 2026
 -   Renamed `Chinese Whispers` to `Label Propagation`.